+++ zingolib/src/wallet/transactions.rs
@@ -1,15 +1,13 @@
 use std::collections::HashMap;
 
-use zcash_primitives::transaction::TxId;
-
-use super::data::{TransactionRecord, WitnessTrees};
+use crate::wallet::{data::WitnessTrees, transaction_records_by_id::TransactionRecordsById};
 
 /// HashMap of all transactions in a wallet, keyed by txid.
 /// Note that the parent is expected to hold a RwLock, so we will assume that all accesses to
 /// this struct are threadsafe/locked properly.
 pub struct TxMapAndMaybeTrees {
-    pub current: HashMap<TxId, TransactionRecord>,
-    pub witness_trees: Option<WitnessTrees>,
+    pub transaction_records_by_id: TransactionRecordsById,
+    witness_trees: Option<WitnessTrees>,
 }
 
 pub mod get;
@@ -19,18 +17,24 @@ pub mod recording;
 impl TxMapAndMaybeTrees {
     pub(crate) fn new_with_witness_trees() -> TxMapAndMaybeTrees {
         Self {
-            current: HashMap::default(),
+            transaction_records_by_id: TransactionRecordsById(HashMap::new()),
             witness_trees: Some(WitnessTrees::default()),
         }
     }
     pub(crate) fn new_treeless() -> TxMapAndMaybeTrees {
         Self {
-            current: HashMap::default(),
+            transaction_records_by_id: TransactionRecordsById(HashMap::new()),
             witness_trees: None,
         }
     }
+    pub fn witness_trees(&self) -> Option<&WitnessTrees> {
+        self.witness_trees.as_ref()
+    }
+    pub(crate) fn witness_trees_mut(&mut self) -> Option<&mut WitnessTrees> {
+        self.witness_trees.as_mut()
+    }
     pub fn clear(&mut self) {
-        self.current.clear();
+        self.transaction_records_by_id.clear();
         self.witness_trees.as_mut().map(WitnessTrees::clear);
     }
 }